Move key reentry coverage out of large files
